# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrailService API",
    "description": "Microservice for managing hiking trails in a wellbeing application",
    "version": "1.0.0",
    "x-logo": {
      "url": "https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png",
      "altText": "TrailService Logo"
    }
  },
  "paths": {
...
    "/api/v1/trails": {
      "get": {
        "tags": [
          "Trails"
        ],
        "summary": "Get Trails",
        "description": "Get all public trails with optional filtering\n# Public users: see only public trails\n# Authenticated users: see public trails + their own private trails",
        "operationId": "get_trails_api_v1_trails_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Difficulty"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by difficulty",
              "title": "Difficulty"
            },
            "description": "Filter by difficulty"
          },
          {
            "name": "min_length",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum trail length in km",
              "title": "Min Length"
            },
            "description": "Minimum trail length in km"
          },
          {
            "name": "max_length",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum trail length in km",
              "title": "Max Length"
            },
            "description": "Maximum trail length in km"
          },
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by location",
              "title": "Location Id"
            },
            "description": "Filter by location"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of trails to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Number of trails to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Trails Api V1 Trails Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error occurred"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trails"
        ],
        "summary": "Create Trail",
        "description": "Create a new trail",
        "operationId": "create_trail_api_v1_trails_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Trail Data"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Trail Api V1 Trails Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid or expired token"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error occurred"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trails/{trail_id}": {
      "get": {
        "tags": [
          "Trails"
        ],
        "summary": "Get Trail",
        "description": "Get detailed information about a specific trail\n# Public: can view public trails only\n# Authenticated: can view public trails + own private trails",
        "operationId": "get_trail_api_v1_trails__trail_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trail_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trail Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Trail Api V1 Trails  Trail Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error occurred"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Trails"
        ],
        "summary": "Update Trail",
        "description": "Update an existing trail",
        "operationId": "update_trail_api_v1_trails__trail_id__put",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "trail_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trail Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Trail Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Trail Api V1 Trails  Trail Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid or expired token"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error occurred"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Trails"
        ],
        "summary": "Delete Trail",
        "description": "Delete a trail",
        "operationId": "delete_trail_api_v1_trails__trail_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "trail_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Trail Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid or expired token"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal server error occurred"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Difficulty": {
        "type": "string",
        "enum": [
          "Easy",
          "Moderate",
          "Hard"
        ],
        "title": "Difficulty"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter session token obtained from authentication"
      }
    }
  },
...
}